requirements.txt
google-cloud-speech==2.26.0
google-cloud-storage==2.16.0
google-api-core==2.18.0
google-auth==2.29.0
python-dotenv==1.0.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
aiofiles==24.1.0
jinja2==3.1.4
